Address CR comment
