+++ src/Windows/Host/Broker/Impl/Services/WindowsSystemInfoService.cs
@@ -1,8 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using System;
 using System.Diagnostics;
+using System.Linq;
 using System.Management;
+using System.Net.NetworkInformation;
 using System.Threading;
 
 namespace Microsoft.R.Host.Broker.Services {
@@ -24,6 +27,40 @@ namespace Microsoft.R.Host.Broker.Services {
             }
 
             return counter / (iterations * 100);
+        }
+
+        public double GetNetworkLoad() {
+            if (!NetworkInterface.GetIsNetworkAvailable()) {
+                return 0;
+            }
+
+            // Select compatible active network interfaces
+            var interfaces = NetworkInterface.GetAllNetworkInterfaces()
+                .Where(x => x.Speed > 0 
+                    && x.Supports(NetworkInterfaceComponent.IPv4) 
+                    && x.Supports(NetworkInterfaceComponent.IPv6) 
+                    && x.OperationalStatus == OperationalStatus.Up 
+                    && x.Speed > 0 && IsCompatibleInterface(x.NetworkInterfaceType)).ToArray();
+
+            // Take initial measurement, wait 500ms, take another.
+            var stats = interfaces.Select(s => s.GetIPStatistics()).ToArray();
+            var initialBytes = stats.Select(s => s.BytesSent + s.BytesReceived).ToArray();
+
+            Thread.Sleep(500);
+
+            stats = interfaces.Select(s => s.GetIPStatistics()).ToArray();
+            var currentBytes = stats.Select(s => s.BytesSent + s.BytesReceived).ToArray();
+
+            // Figure out how many bytes were sent and received within 500ms
+            // and calculate adapter load depending on speed. Take the highest value.
+            double maxLoad = 0;
+            for (int i = 0; i < initialBytes.Length; i++) {
+                // 16 = 8 bits per byte in 1/2 second. Speed it in bits per second.
+                var load = 16.0 * (currentBytes[i] - initialBytes[i]) / interfaces[i].Speed;
+                maxLoad = Math.Max(maxLoad, load);
+            }
+
+            return maxLoad;
         }
 
         public (long TotalVirtualMemory, long FreeVirtualMemory, long TotalPhysicalMemory, long FreePhysicalMemory) GetMemoryInformation() {
@@ -60,6 +97,17 @@ namespace Microsoft.R.Host.Broker.Services {
             return value == null 
                 ? 0
                 : (int.TryParse(value.ToString(), out int result) ? result / divider : 0);
+        }
+
+        private static bool IsCompatibleInterface(NetworkInterfaceType nit) {
+            switch (nit) {
+                case NetworkInterfaceType.Loopback:
+                case NetworkInterfaceType.HighPerformanceSerialBus:
+                case NetworkInterfaceType.Ppp:
+                    return false;
+                default:
+                    return true;
+            }
         }
     }
 }
